use wii remote model

+++ src/model.h
@@ -6,6 +6,6 @@
 extern u32 vertCount;
 extern u32 indexCount;
 extern f32 vertPositions[];
-extern u8 vertIndices[];
+extern u16 vertIndices[];
 
 #endif // MODEL_H
